+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.5.1 (unreleased)
 
+- Improved performance of HNSW index builds
 - Added check for MVCC-compliant snapshot for HNSW index scans
 - Improved performance of index scans for IVFFlat after updates and deletes